Time: 203 ms (65.21%), Space: 107.3 MB (59.25%) - LeetHub

+++ 1161-maximum-level-sum-of-a-binary-tree/1161-maximum-level-sum-of-a-binary-tree.cpp
@@ -21,7 +21,7 @@ public:
         while(!q.empty())
         {
             int size = q.size();
-            long long add = 0;
+            int add = 0;
             level++;
             for(int i = 0; i<size; i++)
             {
